still have cholesky issue, giving up

+++ pinot/data/utils.py
@@ -37,7 +37,7 @@ def from_csv(path, toolkit="rdkit", smiles_col=-1, y_cols=[-2], seed=2666):
             ]
             gs = [pinot.graph.from_oemol(mol) for mol in mols]
 
-        ds = list(zip(gs, list(torch.tensor(df_y.values))))
+        ds = list(zip(gs, list(torch.tensor(df_y.values, dtype=torch.float32))))
         random.seed(seed)
         random.shuffle(ds)
